fix node sync block
fix marshal block when verify sign.
improved tx process.
